+++ operators/ARMORED_classic_mirror.py
@@ -1,4 +1,4 @@
-version = (1, 5, 0)
+version = (1, 6, 0)
 
 import bpy
 
@@ -17,36 +17,42 @@ class MESH_OT_armored_classic_mirror(bpy.types.Operator):
 		return context.active_object is not None and context.active_object.type == 'MESH'
 
 	def execute(self, context):
-		active_object = context.active_object
+		original_active_object = context.active_object
+		selected_objects = [obj for obj in context.selected_objects if obj.type == 'MESH']
 
 		last_mode = context.mode
 		if last_mode == 'EDIT_MESH':
 			bpy.ops.object.mode_set(mode='OBJECT')
+		
+		for obj in selected_objects:
+			mod = obj.modifiers.get('Mirror')
 
-		mod = active_object.modifiers.get('Mirror')
+			if mod is None:
+				self._add_mirror_modifier(context, object=obj)
 
-		if mod is None:
-			self._add_mirror_modifier(context, active_object)
+			elif obj.data.users > 1:
+				self.report({'WARNING'}, 'Cannot apply modifiers to multi-user data.')
+				return {'CANCELLED'}
 
-		elif active_object.data.users > 1:
-			self.report({'WARNING'}, 'Cannot apply modifiers to multi-user data.')
-
-			return {'CANCELLED'}
-
-		else:
-			bpy.ops.object.modifier_copy(modifier='Mirror')
-			bpy.ops.object.modifier_apply(modifier='Mirror.001')
+			else:
+				context.view_layer.objects.active = obj
+				bpy.ops.object.modifier_copy(modifier='Mirror')
+				bpy.ops.object.modifier_apply(modifier='Mirror.001')
 
 		if last_mode == 'EDIT_MESH':
 			bpy.ops.object.mode_set(mode='EDIT')
+		
+		context.view_layer.objects.active = original_active_object
 
 		return {'FINISHED'}
 	
 
-	def _add_mirror_modifier(self, context, active_object) -> bpy.types.Modifier:
-		found_subsurf = bool(active_object.modifiers and active_object.modifiers[-1].type == 'SUBSURF')
+	def _add_mirror_modifier(self, context, object) -> bpy.types.Modifier:
+		context.view_layer.objects.active = object
 
-		mod = active_object.modifiers.new(type='MIRROR', name='Mirror')
+		found_subsurf = bool(object.modifiers and object.modifiers[-1].type == 'SUBSURF')
+
+		mod = object.modifiers.new(type='MIRROR', name='Mirror')
 
 		mod.use_axis             = True,  False, False
 		mod.use_bisect_axis      = True,  True,  True
@@ -55,7 +61,7 @@ class MESH_OT_armored_classic_mirror(bpy.types.Operator):
 		mod.use_clip = True
 
 		if found_subsurf:
-			self._move_modifier_up(context, object=active_object, modifier_name=mod.name)
+			self._move_modifier_up(context, object=object, modifier_name=mod.name)
 
 		return mod